Adds strict password generation
When strict password generation is turned on, it will treat the options
as rules: this means that if the uppercase option is turned on, the
resulting password MUST have an uppercase letter inside of it. The same
works with symbols and numbers.

This will allow compliance with some requirements, where passwords must
contain at least one number, one symbol, etc.

However, it must be noted that although turning on strict password
generation seems like it would increase security, it will theoretically
reduce the entropy of the given password.