fix strict password generation (#11)
Previously, there was a bug in the strict password generation that
wouldn't correctly check whether or not the password contained all the
correct characters more than twice in a row. This is now fixed by using
simple recursion as opposed to a trampoline.

This should be prevented in the future with better tests, that have been
added. This wasn't caught before the merge because of an insignificant
amount of runs for these functions, as the odds of the bug occurring
were relatively low. Each test will now be run 500 times when entropy is
a factor.

This also fixes the regex for symbols.